Implemented user page. Did many improvements on search page.

+++ SwiftyCompanion/Sources/Models/School42Service/Data/SkillData.swift
@@ -3,7 +3,7 @@ import Foundation
 extension School42Service {
 	public struct SkillData: Decodable {
 		public var name: String
-		public var level: Float
+		public var level: Double
 		
 		enum CodingKeys: String, CodingKey {
 			case name
